refactor: move TerminalState source files to src/TerminalState dir
